Fixed bug with requests containing binary data in body.
It was identified that due to the request body being converted to a
String object, requests containing binary data in the body would be
malformed upon being processed by the plugin, potentially due to
encoding bugs(s). In order to avoid this issue, since the request body
is not processed at all, the conversion to a String object was replaced,
and the body is preserved in byte arrays.